+++ www/docs/customization/changelog.md
@@ -25,6 +25,20 @@ changelog:
   # Default is empty
   sort: asc
 
+  # Group commits messages by given regex and title.
+  # Order value defines the order of the groups.
+  # Proving no regex means all commits will be grouped under the default group.
+  # Default is no groups.
+  groups:
+    - title: Features
+      regexp: "^.*feat[(\\w)]*:+.*$"
+      order: 0
+    - title: 'Bug fixes'
+      regexp: "^.*fix[(\\w)]*:+.*$"
+      order: 1
+    - title: Others
+      order: 999
+
   filters:
     # Commit messages matching the regexp listed here will be removed from
     # the changelog